[mcserver] cafezinho: remove, downgrade and add mods
This commit removes the following mods:
- Iron Chests
- Iron Furnances
- Gravestones

Adds the new mods:
- What Are They Up To
- You're In Grave Danger
- Emoji Type
- Mythic Mounts
- GeckoLib
- Patchouli
- CoroUtils
- Cloth Config Api
- Architectury Api
- Fabric Api

And fix the dumb mistake of adding an incompatible version of Xaeros
Mod.

I also removed the field "name" of the mods because it was saving the
mods with the content of the field instead of the jar file name.
(another dumb mistake that cost me hours ik)